Move logistic SGL stuff into separate module. Use custom ProgressParallel for CV estimator.
