+++ afqinsight/sgl.py
@@ -4,7 +4,7 @@ import numpy as np
 import warnings
 
 from functools import partial
-from joblib import Parallel, delayed, effective_n_jobs
+from joblib import delayed, effective_n_jobs
 from scipy import sparse
 from scipy.optimize import root_scalar
 from tqdm.auto import tqdm
@@ -16,11 +16,7 @@ from sklearn.base import (
     is_classifier,
     is_regressor,
 )
-from sklearn.linear_model._base import (
-    LinearClassifierMixin,
-    LinearModel,
-    _preprocess_data,
-)
+from sklearn.linear_model._base import LinearModel, _preprocess_data
 from sklearn.linear_model._coordinate_descent import _alpha_grid as _lasso_alpha_grid
 from sklearn.linear_model._coordinate_descent import _path_residuals
 from sklearn.model_selection import check_cv
@@ -35,7 +31,7 @@ from sklearn.utils.validation import (
 )
 
 from .prox import SparseGroupL1, _soft_threshold
-from .utils import check_groups
+from .utils import check_groups, ProgressParallel
 
 __all__ = []
 
@@ -480,214 +476,6 @@ class SGL(SGLBaseEstimator, RegressorMixin, LinearModel):
         check_is_fitted(self, "is_fitted_")
 
         return safe_sparse_dot(X, self.coef_.T, dense_output=True) + self.intercept_
-
-
-@registered
-class LogisticSGL(SGLBaseEstimator, LinearClassifierMixin):
-    """
-    An sklearn compatible sparse group lasso classifier.
-
-    This solves the sparse group lasso [1]_ problem for a feature matrix
-    partitioned into groups using the proximal gradient descent (PGD)
-    algorithm.
-
-    Parameters
-    ----------
-    l1_ratio : float, default=1.0
-        Hyper-parameter : Combination between group lasso and lasso. l1_ratio=0
-        gives the group lasso and l1_ratio=1 gives the lasso.
-
-    alpha : float, default=0.0
-        Hyper-parameter : overall regularization strength.
-
-    groups : list of numpy.ndarray
-        list of arrays of non-overlapping indices for each group. For
-        example, if nine features are grouped into equal contiguous groups of
-        three, then groups would be ``[array([0, 1, 2]), array([3, 4, 5]),
-        array([6, 7, 8])]``. If the feature matrix contains a bias or
-        intercept feature, do not include it as a group. If None, all
-        features will belong to one group. We set groups in ``__init__`` so
-        that it can be reused in model selection and CV routines.
-
-    scale_l2_by : ["group_length", None], default="group_length"
-        Scaling technique for the group-wise L2 penalty.
-        By default, ``scale_l2_by="group_length`` and the L2 penalty is
-        scaled by the square root of the group length so that each variable
-        has the same effect on the penalty. This may not be appropriate for
-        one-hot encoded features and ``scale_l2_by=None`` would be more
-        appropriate for that case. ``scale_l2_by=None`` will also reproduce
-        ElasticNet results when all features belong to one group.
-
-    fit_intercept : bool, default=True
-        Specifies if a constant (a.k.a. bias or intercept) should be
-        added to the linear predictor (X @ coef + intercept).
-
-    max_iter : int, default=1000
-        Maximum number of iterations for PGD solver.
-
-    tol : float, default=1e-7
-        Stopping criterion. Convergence tolerance for PGD algorithm.
-
-    warm_start : bool, default=False
-        If set to ``True``, reuse the solution of the previous call to ``fit``
-        as initialization for ``coef_`` and ``intercept_``.
-
-    verbose : int, default=0
-        Verbosity flag for PGD solver. Any positive integer will produce
-        verbose output
-
-    suppress_solver_warnings : bool, default=True
-        If True, suppress convergence warnings from PGD solver.
-        This is useful for hyperparameter tuning when some combinations
-        of hyperparameters may not converge.
-
-    Attributes
-    ----------
-    classes_ : ndarray of shape (n_classes, )
-        A list of class labels known to the classifier.
-
-    coef_ : array of shape (n_features,)
-        Estimated coefficients for the linear predictor (`X @ coef_ +
-        intercept_`).
-
-    intercept_ : float
-        Intercept (a.k.a. bias) added to linear predictor.
-
-    n_iter_ : int
-        Actual number of iterations used in the solver.
-
-    Examples
-    --------
-
-    References
-    ----------
-    .. [1]  Noah Simon, Jerome Friedman, Trevor Hastie & Robert Tibshirani,
-        "A Sparse-Group Lasso," Journal of Computational and Graphical
-        Statistics, vol. 22:2, pp. 231-245, 2012
-        DOI: 10.1080/10618600.2012.681250
-
-    """
-
-    def fit(self, X, y):  # pylint: disable=arguments-differ
-        """Fit a linear model using the sparse group lasso
-
-        Parameters
-        ----------
-        X : {array-like, sparse matrix}, shape (n_samples, n_features)
-            The training input samples.
-
-        y : array-like, shape (n_samples,)
-            The target values (class labels in classification, real numbers in
-            regression).
-
-        Returns
-        -------
-        self : object
-            Returns self.
-        """
-        return super().fit(X=X, y=y, loss="log")
-
-    def decision_function(self, X):
-        """
-        Predict confidence scores for samples.
-
-        The confidence score for a sample is the signed distance of that
-        sample to the hyperplane.
-
-        Parameters
-        ----------
-        X : array-like or sparse matrix, shape (n_samples, n_features)
-            Samples.
-
-        Returns
-        -------
-        array, shape=(n_samples,) if n_classes == 2 else (n_samples, n_classes)
-            Confidence scores per (sample, class) combination. In the binary
-            case, confidence score for self.classes_[1] where >0 means this
-            class would be predicted.
-        """
-        check_is_fitted(self)
-
-        X = check_array(X, accept_sparse="csr")
-
-        n_features = self.coef_.size
-        if X.shape[1] != n_features:
-            raise ValueError(
-                "X has %d features per sample; expecting %d" % (X.shape[1], n_features)
-            )
-
-        scores = safe_sparse_dot(X, self.coef_.T, dense_output=True) + self.intercept_
-        return scores
-
-    def predict(self, X):
-        """
-        Predict class labels for samples in X.
-
-        Parameters
-        ----------
-        X : array_like or sparse matrix, shape (n_samples, n_features)
-            Samples.
-
-        Returns
-        -------
-        C : array, shape [n_samples]
-            Predicted class label per sample.
-        """
-        scores = self.decision_function(X)
-        if len(scores.shape) == 1:
-            indices = (scores > 0).astype(np.int)
-        else:
-            indices = scores.argmax(axis=1)
-
-        return self.classes_[indices]
-
-    def predict_proba(self, X):
-        """
-        Probability estimates.
-
-        The returned estimates for all classes are ordered by the label of classes.
-
-        Else use a one-vs-rest approach, i.e calculate the probability of
-        each class assuming it to be positive using the logistic function.
-        and normalize these values across all the classes.
-
-        Parameters
-        ----------
-        X : array-like of shape (n_samples, n_features)
-            Vector to be scored, where `n_samples` is the number of samples and
-            `n_features` is the number of features.
-
-        Returns
-        -------
-        T : array-like of shape (n_samples, n_classes)
-            Returns the probability of the sample for each class in the model,
-            where classes are ordered as they are in ``self.classes_``.
-        """
-        check_is_fitted(self)
-        return super()._predict_proba_lr(X)
-
-    def predict_log_proba(self, X):
-        """
-        Predict logarithm of probability estimates.
-
-        The returned estimates for all classes are ordered by the label of classes.
-
-        Parameters
-        ----------
-        X : array-like of shape (n_samples, n_features)
-            Vector to be scored, where `n_samples` is the number of samples and
-            `n_features` is the number of features.
-
-        Returns
-        -------
-        T : array-like of shape (n_samples, n_classes)
-            Returns the log-probability of the sample for each class in the
-            model, where classes are ordered as they are in ``self.classes_``.
-        """
-        return np.log(self.predict_proba(X))
-
-    def _more_tags(self):  # pylint: disable=no-self-use
-        return {"binary_only": True, "requires_y": True}
 
 
 def _alpha_grid(
@@ -1378,10 +1166,8 @@ class SGLCV(LinearModel, RegressorMixin, TransformerMixin):
                 X_order="F",
                 dtype=X.dtype.type,
             )
-            for this_l1_ratio, this_alphas in zip(
-                tqdm(l1_ratios, desc="L1_ratios", total=n_l1_ratio), alphas
-            )
-            for train, test in tqdm(folds, desc="CV folds", total=len(folds))
+            for this_l1_ratio, this_alphas in zip(l1_ratios, alphas)
+            for train, test in folds
         )
 
         if isinstance(self.verbose, int):
@@ -1391,9 +1177,12 @@ class SGLCV(LinearModel, RegressorMixin, TransformerMixin):
         else:
             parallel_verbosity = self.verbose
 
-        mse_paths = Parallel(
+        mse_paths = ProgressParallel(
             n_jobs=self.n_jobs,
             verbose=parallel_verbosity,
+            use_tqdm=bool(self.verbose),
+            desc="L1 ratios * CV folds",
+            total=n_l1_ratio * len(folds),
             **_joblib_parallel_args(prefer="threads")
         )(jobs)
