+++ afqinsight/utils.py
@@ -2,7 +2,10 @@ from __future__ import absolute_import, division, print_function
 
 import numpy as np
 import matplotlib.pyplot as plt
+
 from collections import namedtuple
+from joblib import Parallel
+from tqdm.auto import tqdm
 
 __all__ = []
 
@@ -135,3 +138,23 @@ def check_groups(groups, X, allow_overlap=False, fit_intercept=True):
             raise ValueError("Overlapping groups detected.")
 
     return tuple(groups)
+
+
+class ProgressParallel(Parallel):
+    def __init__(self, use_tqdm=True, total=None, desc=None, *args, **kwargs):
+        self._use_tqdm = use_tqdm
+        self._total = total
+        self._desc = desc
+        super().__init__(*args, **kwargs)
+
+    def __call__(self, *args, **kwargs):
+        with tqdm(
+            disable=not self._use_tqdm, desc=self._desc, total=self._total
+        ) as self._pbar:
+            return Parallel.__call__(self, *args, **kwargs)
+
+    def print_progress(self):
+        if self._total is None:
+            self._pbar.total = self.n_dispatched_tasks
+        self._pbar.n = self.n_completed_tasks
+        self._pbar.refresh()
